Refactor Makefile, fix install target overriding DESTDIR
The initial change was because DESTDIR was used incorrectly to the
standard way of it's usual convention of it being the variable
to set the destination directory of the prefix but not the final
directory.

Further I changed the install target to use install instead of
mkdir and cp wildcard as it's simpler and cleaner.

I also removed the default clean target as it removes all elc files
when trying to compile the bytecode again by default which could also
be done by using make -B if so desired.

The autoloads file is no longer a phony target as it now correctly
depends on all the Emacs Lisp source files.

Signed-off-by: Björn Bidar <bjorn.bidar@thaodan.de>